Update Dependencies ⭐

+++ src/capture.rs
@@ -152,7 +152,7 @@ impl<T: GraphicsCaptureApiHandler + Send + 'static, E> CaptureControl<T, E> {
 
         if let Some(thread_handle) = self.thread_handle.take() {
             let handle = thread_handle.as_raw_handle();
-            let handle = HANDLE(handle as isize);
+            let handle = HANDLE(handle);
             let therad_id = unsafe { GetThreadId(handle) };
 
             loop {
